refactor view package
