+++ src/main/java/com/codecool/view/windowedview/WindowedView.java
@@ -1,0 +1,10 @@
+package com.codecool.view.windowedview;
+
+import com.codecool.model.board.Board;
+import com.codecool.view.View;
+
+public class WindowedView implements View {
+    public void draw(Board board) {
+        
+    }
+}
